before delay

+++ shared/src/commonMain/kotlin/com/presta/customer/ui/components/addSavings/DefaultAddSavingsComponent.kt
@@ -113,7 +113,7 @@ class DefaultAddSavingsComponent(
                             phoneNumber = state.cachedMemberData.phoneNumber,
                             loanRefId = null,
                             beneficiaryPhoneNumber = null,
-                            amount = amount.toInt(),
+                            amount = amount,
                             paymentType = mode
                         )
                     )
